Video 12 Done

+++ TweetBook-NetCore-REST API/Controllers/V1/PostsController.cs
@@ -9,6 +9,7 @@ using TweetBook.Domain;
 using TweetBook_NetCore_REST_API.Contracts.Requests;
 using TweetBook_NetCore_REST_API.Contracts.Responses;
 using TweetBook_NetCore_REST_API.Contracts.V1;
+using TweetBook_NetCore_REST_API.Extensions;
 using TweetBook_NetCore_REST_API.Services;
 
 namespace TweetBook.Controllers.V1
@@ -43,11 +44,13 @@ namespace TweetBook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute]Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post
-            {
-                Id = postId,
-                Name = request.Name,
-            };
+            var userOwnPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnPost)
+                return BadRequest(new { error = "You do not own this post" });
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
 
             var updated = await _postService.UpdatePostAsync(post);
 
@@ -60,7 +63,11 @@ namespace TweetBook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post
+            {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
             //var post = new Post { Id = postRequest == null ? Guid.NewGuid() : postRequest.Id };
 
             await _postService.CreatePostAsync(post);
@@ -75,6 +82,11 @@ namespace TweetBook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute]Guid postId)
         {
+            var userOwnPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnPost)
+                return BadRequest(new { error = "You do not own this post" });
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
